requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
